refactor: modify structure according to new setup
