+++ vae/__init__.py
@@ -1,5 +1,7 @@
 from vae import problem
 from vae import datastream
 from vae import architecture
+from vae import metrics
 
+from vae.log_examples import log_examples
 from vae.train import train
