+++ vae/datastream/__init__.py
@@ -1,3 +1,4 @@
+from vae.datastream.datasets import datasets
 from vae.datastream.augmenter import augmenter
 from vae.datastream.evaluate_datastreams import (
     evaluate_datastreams
