+++ vae/datastream/evaluate_datastreams.py
@@ -1,16 +1,11 @@
-import numpy as np
-import pandas as pd
 from datastream import Datastream
 
-from vae import problem, splits
+from vae import datastream
 
 
 def evaluate_datastreams():
-    evaluate_datasets = problem.evaluate_datasets()
+    datasets = datastream.datasets()
     return {
-        split_name: Datastream(dataset)
-        for split_name, dataset in dict(
-            gradient=evaluate_datasets['train'],
-            compare=evaluate_datasets['compare'],
-        ).items()
+        name: Datastream(dataset).take(256)
+        for name, dataset in datastream.datasets().items()
     }
