+++ vae/problem/__init__.py
@@ -1,4 +1,2 @@
 from vae.problem.example import Example
-from vae.problem.evaluate_datasets import (
-    evaluate_datasets
-)
+from vae.problem.datasets import datasets
