--- vae/splits/__init__.py
@@ -1,5 +1,0 @@
-from pathlib import Path
-
-splits = Path(__file__).parent
-
-compare = splits / 'compare.json'
